requirements.txt
pdfplumber
python-docx
natasha
openpyxl
pytest
